Split our generate service into 2
This should make writing the tests much easier.
